src/flashcardWorker/requirements.txt
numpy==1.21.5
pymongo==3.10.1
python-telegram-bot==13.8.1